feat/#1 : 공통 컴포넌트 분리 & 코드 위치 조정
