+++ client-quota/src/components/teams/createForm/CreateGroupLanding.jsx
@@ -1,25 +1,40 @@
 import styled from "styled-components";
 import CreateGroupForm from "./ui/CreateGroupForm";
 
+const CreateGroupLanding = () => {
+    return (
+        <CreateGroupFormContainer>
+            <GroupFormContainer>
+                <GroupFromTitleContainer>
+                    <BackButton onClick={() => window.history.back()}>←</BackButton>
+                    <Title>그룹 추가하기</Title>
+                </GroupFromTitleContainer>
+                <CreateGroupForm/>
+            </GroupFormContainer>
+        </CreateGroupFormContainer>
+    );    
+}
+
+export default CreateGroupLanding;
+
 const CreateGroupFormContainer = styled.div`
     display: flex;
     justify-content: center;
     align-items: center;
     height: 100vh;
-`
+`;
 
 const GroupFormContainer = styled.div`
-`
+`;
 
 const GroupFromTitleContainer = styled.div`
     display: flex;
     align-items: center;
-
-`
+`;
 
 const Title = styled.h1`
    margin-left: 5px;
-`
+`;
 
 const BackButton = styled.button`
     display: flex;
@@ -36,21 +51,4 @@ const BackButton = styled.button`
 
     border: none;
     cursor: pointer;
-`
-
-
-const CreateGroupLanding = () => {
-    return (
-        <CreateGroupFormContainer>
-            <GroupFormContainer>
-                <GroupFromTitleContainer>
-                    <BackButton onClick={() => window.history.back()}>←</BackButton>
-                    <Title>그룹 추가하기</Title>
-                </GroupFromTitleContainer>
-                <CreateGroupForm/>
-            </GroupFormContainer>
-        </CreateGroupFormContainer>
-    );    
-}
-
-export default CreateGroupLanding;
+`;
