+++ client-quota/src/components/teams/createForm/ui/CreateGroupForm.jsx
@@ -1,9 +1,31 @@
 import styled from "styled-components";
-import InputTitle from "./InputTitle";
-import InputText from "./InputText";
+import InputTitle from "../../../common/InputTitle";
+import InputText from "../../../common/InputText";
 import InputProfile from "./InputProfile";
-import InputGroupLink from "./InputGroupLink";
-import InputTextarea from "./InputTextarea";
+import InputLink from "../../../common/InputLink";
+import InputTextarea from "../../../common/InputTextarea";
+
+const CreateGroupForm = () => {
+    return (
+        <StyledGroupForm>
+            <InputTitle>그룹 프로필</InputTitle>
+            <InputProfile name="teamProfileImage"/>
+
+            <InputTitle required>그룹 입력</InputTitle>
+            <InputText name="teamName" placeholder="그룹 이름을 입력해 주세요."/>
+
+            <InputTitle required>그룹 링크</InputTitle>
+            <InputLink name="teamUrl" placeholder="GroupExample">quotime.co.kr/</InputLink>
+
+            <InputTitle>회사/그룹 소개</InputTitle>
+            <InputTextarea name="teamDescription" placeholder="우리 회사/그룹에 대해 설명해 주세요."/>
+
+            <CompleteButton>완료</CompleteButton>
+        </StyledGroupForm>
+    );
+}
+
+export default CreateGroupForm;
 
 const StyledGroupForm = styled.form`
     width: 625px;
@@ -28,26 +50,4 @@ const CompleteButton = styled.button`
     margin: 20px auto 0 auto;
 
     cursor: pointer;
-    
 `
-
-const CreateGroupForm = () => {
-    return <StyledGroupForm>
-        <InputTitle>그룹 프로필</InputTitle>
-        <InputProfile name="teamProfileImage"/>
-
-        <InputTitle required>그룹 입력</InputTitle>
-        <InputText name="teamName" placeholder="그룹 이름을 입력해 주세요."/>
-
-        <InputTitle required>그룹 링크</InputTitle>
-        <InputGroupLink name="teamUrl" placeholder="GroupExample"/>
-
-        <InputTitle>회사/그룹 소개</InputTitle>
-        <InputTextarea name="teamDescription" placeholder="우리 회사/그룹에 대해 설명해 주세요."/>
-
-        <CompleteButton>완료</CompleteButton>
-    </StyledGroupForm>
-    
-}
-
-export default CreateGroupForm;
